pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "auditlog-ai"
version = "0.2.0"
description = "Automatic compliance logging for AI agents. One line install, vendor-agnostic, audit-ready."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
dependencies = []
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
]

[project.optional-dependencies]
supabase = ["supabase>=2.0.0"]
postgres = ["psycopg2-binary>=2.9.0"]
async = ["aiofiles>=23.0.0"]
all = [
    "supabase>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "aiofiles>=23.0.0",
]

[project.urls]
Homepage = "https://github.com/0xqlive/agentlog"
Documentation = "https://github.com/0xqlive/agentlog#readme"
Issues = "https://github.com/0xqlive/agentlog/issues"
